+++ examples/AspNetCore/Startup.cs
@@ -75,7 +75,6 @@ namespace Examples.AspNetCore
                         .AddHttpClientInstrumentation()
                         .AddZipkinExporter(zipkinOptions =>
                         {
-                            zipkinOptions.ServiceName = this.Configuration.GetValue<string>("Zipkin:ServiceName");
                             zipkinOptions.Endpoint = new Uri(this.Configuration.GetValue<string>("Zipkin:Endpoint"));
                         }));
                     break;