+++ src/OpenTelemetry.Exporter.Console/ConsoleActivityExporter.cs
@@ -19,7 +19,7 @@ using System.Diagnostics;
 using System.Linq;
 using OpenTelemetry.Resources;
 
-namespace OpenTelemetry.Exporter.Console
+namespace OpenTelemetry.Exporter
 {
     public class ConsoleActivityExporter : ConsoleExporter<Activity>
     {
@@ -69,15 +69,6 @@ namespace OpenTelemetry.Exporter.Console
                         {
                             this.WriteLine($"        {attribute.Key}: {attribute.Value}");
                         }
-                    }
-                }
-
-                if (activity.Baggage.Any())
-                {
-                    this.WriteLine("Activity.Baggage:");
-                    foreach (var baggage in activity.Baggage)
-                    {
-                        this.WriteLine($"    {baggage.Key}: {baggage.Value}");
                     }
                 }